+++ src/model/loop.ml
@@ -64,7 +64,8 @@ let empty ~mode () = {
 let pp_wrap pp fmt x =
   Format.fprintf fmt "`%a`" pp x
 
-let code = Dolmen_loop.Code.create
+let code =
+  Dolmen_loop.Code.create
     ~category:"Model"
     ~descr:"on model verification errors"